@@ -1,6 +1,6 @@
 module storj.io/common
 
-go 1.20
+go 1.23.0
 
 require (
 	cloud.google.com/go/profiler v0.4.0
