@@ -48,10 +48,7 @@ func Decode(b string) []byte {
 	// Of course, in addition, we'll need to handle boundary condition when `b` is not multiple of 58^10.
 	// In that case we'll use the bigRadix[n] lookup for the appropriate power.
 	for t := b; len(t) > 0; {
-		n := len(t)
-		if n > 10 {
-			n = 10
-		}
+		n := min(len(t), 10)
 
 		total := uint64(0)
 		for k := 0; k < n; k++ {
@@ -121,7 +118,7 @@ func Encode(b []byte) string {
 			}
 		} else {
 			m := mod.Int64()
-			for i := 0; i < 10; i++ {
+			for range 10 {
 				answer = append(answer, alphabet[m%58])
 				m /= 58
 			}
