@@ -59,7 +59,7 @@ func TestBase58Check(t *testing.T) {
 	// case 2: invalid formats (string lengths below 5 mean the version byte and/or the checksum
 	// bytes are missing).
 	testString := ""
-	for len := 0; len < 4; len++ {
+	for range 4 {
 		testString += "x"
 		_, _, err = base58.CheckDecode(testString)
 		if !errors.Is(err, base58.ErrInvalidFormat) {
