@@ -434,8 +434,8 @@ func FindDefaultsParam() string {
 func FindFlagEarly(flagName string) string {
 	// workaround to have early access to 'dir' param
 	for i, arg := range os.Args {
-		if strings.HasPrefix(arg, fmt.Sprintf("--%s=", flagName)) {
-			return strings.TrimPrefix(arg, fmt.Sprintf("--%s=", flagName))
+		if after, ok := strings.CutPrefix(arg, fmt.Sprintf("--%s=", flagName)); ok {
+			return after
 		} else if arg == "--"+flagName && i < len(os.Args)-1 {
 			return os.Args[i+1]
 		}
