@@ -4,6 +4,7 @@
 package rpctest
 
 import (
+	"slices"
 	"sync"
 
 	"storj.io/common/rpc/rpcpool"
@@ -52,7 +53,7 @@ func (r *CallRecorder) RecordCall(rpc string) {
 
 // History returns the list of rpc names which called on this connection.
 func (r *CallRecorder) History() []string {
-	return append([]string{}, r.calls...)
+	return slices.Clone(r.calls)
 }
 
 // Attach wraps a drpc.Conn connection and returns with one where the counters are hooked in.
